[Cashtab] Deprecate ecashjs-lib
Summary:
T2893

This diff deprecates `ecashjs-lib` from Cashtab and replaces it with BitGo's `utxo-lib`. Unfortunately, we are still waiting for Bitgo to publish a version of `utxo-lib` that includes `ecash` transaction support. However, it is probably worth doing this drop-in now with the shims (i.e. the `0x01 | 0x40` in `cashMethods.js`) and then upgrading it later.

Test Plan:
`npm test`

`npm start`
send an XEC tx
send an XEC tx with a cashtab msg
send an XEC tx with an encrypted cashtab msg
send an etoken tx
sign a msg
verify the msg

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D13679