[alias-server] Bump avalancheCheckCount
Summary:
T3060

In testing over 20 blocks, values between 50 and 90 were observed. The reason this param can be set in config is to handle observed behavior of the server. Will need to be adjusted for final prod environment.

Default needs to be higher to handle local testing conditions.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D13685