[e.cash] Add navbar component
Summary: [T3111 ] Adding the navbar component. I seperated out the actual nav items and links to a seperate document for readability and because it will be used for the footer

Test Plan: run the app and see how it looks and performs. It is fully responsive as well so check that out

Reviewers: #bitcoin_abc, bytesofman, Mengerian

Reviewed By: #bitcoin_abc, bytesofman, Mengerian

Subscribers: Mengerian, bytesofman

Differential Revision: https://reviews.bitcoinabc.org/D13672