[ecashaddrjs] Ensure README displays at npmjs page
Summary:
T3119

Not sure what's going on here. README worked for 1.3.0, not for 1.4.0. However if I try reverting to the 1.3.0 README, no good, doesn't work. Maybe something has changed with how the file is linted.

Will come back to this after working on other things and see if I realize what I'm overlooking.

Test Plan: need to publish this to npm to see how the readme is shown on their site. Currently still not working.

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D13703