[chronik mock] Make sure history function is async so Promise.all() tests work
Summary:
Discovered this bug in creating D13707

The chronik function to get one page of tx history is async. For efficiency, getting multiple pages is usually done with `Promise.all()`. This will only work on an array of promises, which need to be async functions.

Test Plan:
See passing unit tests in D13707
Check unit tests for alias-server and ecash-herald here

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D13727