[linter] Work around a shellcheck issue
Summary:
Shellcheck version 0.9.0 is failing on the verify.sh script with the following error:
```
shellcheck: (Array.!): undefined array element
```

The issue has been narrowed down to the $? output check in the if/else branch, and this diff is a fix with no change in behavior.

Test Plan: Check the code is stricly equivalent.

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D13762