[alias spec] Reword specification to make it more descriptive
Summary:
Mostly just more detailed descriptions of the various pieces.
Still needs more work, but this is a step towards making it more complete.

Also changes the spec to having the Alias Address defined in an op-return push rather than the sending "from" address.

Depends on D13736

Test Plan: Read it to see that it all makes sense

Reviewers: bytesofman, emack, #bitcoin_abc

Reviewed By: emack, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D13737