[arclint] Lint .arclint file with prettier
Summary:
T3147

When I modify and save `.arclint` on my local machine, the file is linted by prettier as I have it installed.

These changes seem to be useful, so extending `prettier` to lint beyond the `web/` folder would be nice. At any rate, this needs to happen because `web/` is no longer a catch-all folder for javascript files in the monorepo.

This diff just lints `.arclint`. Will stack a diff to fix the target directories of `prettier` in `.arclint`

Test Plan: Confirm these changes are desired

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D13797