[Chronik] Add `/blocks/:start_height/:end_height` HTTP endpoint
Summary:
Allows querying a long range of blocks by start and end height (inclusive).

The limit of 500 may seem high, but this is a very cheap call with no extra calls to the DB, so this is fine.

Depends on D13829.

Test Plan: `ninja && ./test/functional/test_runner.py chronik_blocks`

Reviewers: Fabien, #bitcoin_abc

Reviewed By: Fabien, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D13830