[Chronik] Add `/block-txs/:hash_or_height` HTTP endpoint
Summary:
Allows querying the txs of a block in paginated fashion.

The NNG Chronik returned these as part of the `Block` protobuf message, but it now is clear that this poses a DoS vector for larger blocks.

Depends on D13857.

Test Plan: `ninja && ./test/functional/test_runner.py chronik_block_txs`

Reviewers: Fabien, #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Subscribers: PiRK

Differential Revision: https://reviews.bitcoinabc.org/D13858