[herald] Patch bug in splitOverflowTgMsg
Summary:
T2972

Patch a bug in splitOverflowTgMsg

The function was resetting the count of the 'next' message to zero. However, the 'next' message will include the line currently being iterated on in the loop. So, length should not be reset to zero but to the current length of that line.

In most cases the function will still work. Error cases were discovered when adding token send parsing changed msg lengths.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D13865