kernel: Add headerTip method to notifications
Summary:
This commit is part of the libbitcoinkernel project and seeks to remove
the ChainstateManager's and, more generally, the kernel library's
dependency on interface_ui with options methods in this and the following
few commits. By removing interface_ui from the kernel library, its
dependency on boost is reduced to just boost::multi_index.

This is a partial backport of [[https://github.com/bitcoin/bitcoin/pull/27636 | core#27636]]
https://github.com/bitcoin/bitcoin/pull/27636/commits/84d71457e7250ab25c0a11d1ad1c7657197ffd90

Depends on D16204

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D16205