Add a checkpoint after the May 15th, 2023 upgrade
Summary:
The upgrade has activated, add a checkpoint so that people that missed the upgrade can sync the chain more easily.

The actual upgrade blocks are 792,117 (mainnet) and 1,556,118 (testnet).
The checkpointed blocks are the first ones containing transactions after the rules are activated, to enforce the replay protection.

Test Plan: Run IBD

Reviewers: #bitcoin_abc, Mengerian, Fabien

Reviewed By: #bitcoin_abc, Mengerian, Fabien

Subscribers: Mengerian

Differential Revision: https://reviews.bitcoinabc.org/D13897