Add a test-only createwallettransaction RPC
Summary:
This makes it possible to create a wallet transaction without broadcasting it. Thi swill be usefule for testing after we remive the pre-wellington chained tx-limit as it becomes impossible to generate a wallet that is rejected by the mempool anymore.

Because the tx is now valid upon reorg, the test has to be changed a bit.

Test Plan:
  ./test/functional/test_runner.py wallet_balance.py

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Subscribers: PiRK

Differential Revision: https://reviews.bitcoinabc.org/D13929