chainparams: add blockhash to AssumeutxoData
Summary:
This allows us to reference assumeutxo configuration by blockhash as
well as height; this is helpful in future changes when we want to
reference assumeutxo configurations before the block index is loaded.

Also remove the unused assumeutxo checkpoint for block 210.

This is a partial backport of [[https://github.com/bitcoin/bitcoin/pull/27596 | core#27596]]

https://github.com/bitcoin/bitcoin/pull/27596/commits/434495a8c1496ca23fe35b84499f3daf668d76b8

Depends on D17633

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, roqqit

Reviewed By: #bitcoin_abc, roqqit

Differential Revision: https://reviews.bitcoinabc.org/D17634