[e.cash] Add faq page
Summary:
Adding the faq page. Created list of faqs and rendering them on the page. Note: made a style tweak to the H2 component
as the long title revealed a flaw

Test Plan:
preview the site and navigate to the faq page
check everything looks good on both themes and all screen sizes
check content for any mistakes

Reviewers: #bitcoin_abc, bytesofman

Reviewed By: #bitcoin_abc, bytesofman

Subscribers: bytesofman

Differential Revision: https://reviews.bitcoinabc.org/D13938