[Chronik] Add `Node`, a wrapper struct to access the node from Chronik
Summary:
In order to broadcast txs from Chronik, we need to be able to talk to the node via `ChronikBridge`.

For this, we add `Node`, which contains a `UniquePtr` to the bridge, and add it as an `Extension` layer to the HTTP router. This way we can access it from HTTP.

Also change `Chronik` to use `Node`, so we can re-use the same struct.

Test Plan: `ninja check-crates`

Reviewers: Fabien, #bitcoin_abc

Reviewed By: Fabien, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D13987