[ecash-herald] Prep OP_RETURN strings for tg parsing depending on content
Summary:
T2972

See D14030. This change is needed to support more complicated parsing of OP_RETURN msgs, i.e. if some of the data could be used to generate a link in the tg msg.

Currently, ecash-herald calls `prepareStringForTelegramHTML` on the whole `msg` for OP_RETURN txs. However, we want this to depend on the msg contents.

At this point the app doesn't support any complicated OP_RETURN pushes so it's more or less a non-change. D14030 shows how this can be used to create better msgs for OP_RETURN txs.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D14031