[secp256k1] Normalize ge produced from secp256k1_pubkey_load
Summary:
The output ge is normalized when sizeof(secp256k1_ge_storage) = 64
but not when it's not 64. ARG_CHECK at the end of the function
assumes normalization. So normalize ge in the other code path too.

This is a backport of [[https://github.com/bitcoin-core/secp256k1/pull/1349 | secp256k1#1349]]

Test Plan: `ninja check-secp256k1`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D20008