[alias-server] Bring generateMocks script up to date with app
Summary:
T3060

This script was not updated in line with the rest of the app, as mocks were not changed until the most recent spec was implemented.

Test Plan: `npm run generateMocks` (FYI takes about 30 seconds, as the tx history of the IFP address is quite large, ~130 MB. Note when the app is in prod, this only happens on startup with an empty db).

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Subscribers: emack

Differential Revision: https://reviews.bitcoinabc.org/D14074