[alias-server] Ignore tx history before aliases active
Summary:
T3060

chronik can be overloaded on app startup if it fetches the entire history of the IFP address. We have no need to fetch the tx history before alias txs start. So, initialize server state to that point.

Note: this can be deprecated when we migrate to the node version of chronik, which allows getting txHistory after a set txid.

Test Plan:
`npm test`

If you want to confirm there are in fact 45587 confirmed txs thru blockheight 785000 on the IFP address, can do with by

`npm run generateMocks`
Inspect `test/mocks/generated/aliasTxHistory.json` (will need to write a script, it's an array ~130MB)

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Subscribers: emack

Differential Revision: https://reviews.bitcoinabc.org/D14093