[Apps] [Examples] Initial mocha framework and getDetailsFromTxid example
Summary:
T3203

This diff creates an Example folder under Apps on the monorepo and which will be populated with example code as a reference guide for new devs looking into building on eCash.

This is an initial diff containing the readme and the getDetailsFromTxid example along with the mocha test suite initialized.

Test Plan:
`nvm install 16`
`npm i`
`npm run getDetailsFromTxid  <chronik url><txid>` and ensure details for the txid is displayed and matches the block explorer
`npm test` and ensure all tests pass

Reviewers: bytesofman, #bitcoin_abc, Fabien

Reviewed By: bytesofman, #bitcoin_abc, Fabien

Subscribers: tobias_ruck, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D14100